requirements.txt
django-admin-log==0.2
-e git://github.com/thornomad/django-hitcount.git@602038c445a55db4b9d33a9c1b6929c09539d42f#egg=django_hitcount-master
-e git://github.com/webstack/django-endless-pagination.git@0ed7376fa461345d96ce62541816dd3550bb0d25#egg=django_endless_pagination